casted to force expected(?) branching and index wrap. addresses coverity run 8 CID 330
svn:revision:48818
svn:branch:trunk
svn:account:indianlarry

+++ src/sig/syn.c
@@ -81,8 +81,8 @@ Usage: syn samples_per_set [ratio] > doubles\n";
 	    if (ret != 1)
 		perror("fwrite");
 	    findex += stepsize;
-	    if (findex > TABSIZE)
-		findex -= TABSIZE;
+	    if ((int)findex > TABSIZE)
+		findex -= (double)TABSIZE;
 	}
     }
     return 0;
